refactor: trait cell created to invert dependencies, Cell renamed to FinitePlaneCell
