+++ src/domain/cells/mod.rs
@@ -1,0 +1,1 @@
+pub mod finite_plane_cell;
