+++ src/domain/mod.rs
@@ -1,2 +1,3 @@
-pub mod cell;
+pub mod cells;
 pub mod grid;
+pub mod ports;
