+++ src/tests/domain/cells/mod.rs
@@ -1,0 +1,1 @@
+pub mod test_euclidian_cell;
