+++ src/tests/domain/mod.rs
@@ -1,2 +1,3 @@
 pub mod test_cell;
 pub mod grid;
+pub mod cells;
